Updated HotBar to Use SUIButtonDisplay
There are two lines of bad code that I could look into resolving, but the work may be more than its worth.

HotBar switched to SUIButtonDisplay as it is planned to support two rows of selectable icon, with the bottom being the 'main' selectable and the top being defined by the bottom's selection. I.e. Cheat Sheet's Paint Tool UI.